% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/src_spark_sql.R
\name{sql_join.Hive2Connection}
\alias{sql_join.Hive2Connection}
\title{#}
\usage{
\method{sql_join}{Hive2Connection}(con, x, y, type = "inner", by = NULL,
  ...)
}
\description{
#
}
\references{
\url{https://github.com/wush978/dplyrimpaladb/blob/dplyrSparkSQL-reference/R/src-impaladb.r#L323}
}

